Refactoring Chatbot
