+++ terraform/modules/agentic-chatbot/main.tf
@@ -112,8 +112,7 @@ resource "null_resource" "docker_push" {
     docker_file = filemd5("${path.module}/../../../apps/chatbot/Dockerfile")
     app = filemd5("${path.module}/../../../apps/chatbot/app.py")
     requirements = filemd5("${path.module}/../../../apps/chatbot/requirements.txt")
-    bedrock_cliebt = filemd5("${path.module}/../../../apps/chatbot/bedrock_client.py")
-    embedder = filemd5("${path.module}/../../../apps/chatbot/embedder.py")
+    llm_client = filemd5("${path.module}/../../../apps/chatbot/llm_client.py")
     kubernetes_resource = filemd5("${path.module}/../../../apps/chatbot/kubernetes_resource.py")
     retriever = filemd5("${path.module}/../../../apps/chatbot/retriever.py")
     logger = filemd5("${path.module}/../../../apps/chatbot/logger.py")
